Refactor type handling in StdControl.h and CStrictInteger.h for improved safety and compatibility
- Updated type checks in StdControl.h to use static_cast for buffer size comparisons, enhancing type safety.
- Introduced new macro in CStrictInteger.h to handle comparisons with ssize_t when it is the same as int, preventing potential double definition issues.
- Improved overall type consistency across the codebase to support better data handling and compatibility with 64-bit architecture.

+++ sakura_core/apiwrap/StdControl.h
@@ -84,9 +84,9 @@ namespace ApiWrap{
 		const int bufsize = length + 1;
 		
 		// ウィンドウタイトルを設定するのに必要なバッファを確保する
-		if (str.capacity() < bufsize)
+		if (str.capacity() < static_cast<size_t>(bufsize))
 		{
-			str.AllocStringBuffer(bufsize);
+			str.AllocStringBuffer(static_cast<size_t>(bufsize));
 		}
 
 		int actualCount = ::GetWindowText(hwnd, str.GetStringPtr(), static_cast<int>(str.capacity()));
@@ -103,7 +103,7 @@ namespace ApiWrap{
 				return false;
 			}
 		}
-		else if(actualCount >= str.capacity())
+		else if(static_cast<size_t>(actualCount) >= str.capacity())
 		{
 			// GetWindowText() の仕様上はありえないはず
 			return false;
@@ -186,7 +186,7 @@ namespace ApiWrap{
 		{
 			return false;
 		}
-		else if(str.capacity() <= actualCount)
+		else if(str.capacity() <= static_cast<size_t>(actualCount))
 		{
 			return false;
 		}
